==> scripts/00-install_packages.R <==
#### Preamble ####
# Purpose: Install necessary R packages needed to run scripts
# Author: Adrian Ly
# Date: 18 March 2024
# Contact: adrian.ly@mail.utoronto.ca
# License: MIT

#### Workspace setup ####

install.packages("readr")
install.packages("lubridate")
install.packages("tidyverse")
install.packages("dplyr")
install.packages("tidyr")
install.packages("knitr")
install.packages("janitor")
install.packages("scales")
install.packages("RColorBrewer")
install.packages("ggplot2")
install.packages("kableExtra")
install.packages("here")
install.packages("arrow")
install.packages("rstanarm")
install.packages("modelsummary")
install.packages("lme4")
